(PUP-9717) Improve exception handling responsibilities
Previously some states raised an exception if they failed to transition to the
next state, while others returned a Wait object. This made it harder to reason
about control flow since responsibility was shared among the states and machine.
For example, some states would raise if the server wasn't running yet, while
other states would wait and retry.

This commit simplifies the contract between states and the state machine. If a
state handles an exception, then it should return the next state to transition
to. Typically it will return an Error state using the `to_error` helper.
However, it can also be a Wait state, in the case where the CSR hasn't been
signed yet, or a NeedCert state, in the case where the server already has a CSR
with that name.

If a state doesn't choose to handle an exception, then the state machine will
rescue it, wrap it in an Error state, and for now, will re-raise the original
exception.

One outcome of this change is it's easier to test when an Error occurs, since we
can inspect its message directly and not have to inspect the logs.