(PUP-9717) Don't wrap exceptions unnecessarily
Previously, the state machine created a Puppet::Error object whose message was
the same as its cause's message. However, the output is redundant due to the way
Puppet.log_exception unwraps exceptions.

    Error: Failed to retrieve certificate for foobar: Not Found
    Wrapped exception:
    Not Found

Instead create a Puppet::Error object with the same message and backtrace as its
cause, but don't preserve the cause.