(PUP-9720) Create pending test fingerprint fallback behavior
Added fingerprint tests for falling back to the request, falling back to neither
and an exception is raised. The code was updated to not call `nil.client_cert`
as that was causing the pending tests to pass.

Memoize a `machine` and use that in tests.

Reduce the scope of what 'should not wait for a certificate in fingerprint mode'
was expecting. Move the other `expects` to a different test.