(PUP-9720) Fallback to CSR fingerprint if cert doesn't exist
This restores the pre-6.4 behavior of printing the CSR fingerprint if the client
cert doesn't exist. It only looks in the local file system and doesn't change
system state as one would expect when running `puppet agent --fingerprint`.

It does not restore the behavior of downloading the CA/CRL bundle, generating a
key (if missing) or attempting to download the client cert.

If an exception occurs, such as the client cert is invalid, then the exception
is logged instead of using writing to stderr so that the backtrace can be
displayed when running with --trace.