(PUP-9794) prevent autoload superclass errors
With this commit, 'package_targetable.rb' requires 'puppet/provider/package'
avoiding errors when the gem package provider is the first provider to
be autoloaded, which varies depending on the inode order of files in
'puppet/provider/package' and which one 'Dir.entries' happens to returns first.

The 'package_provider' fact in puppetlabs-stdlib triggers this error when it
requires the 'puppet/type/package' outside the body of 'setcode'.

In between the time that 'puppet/type.rb' is required and when when providers
are loaded 'Puppet::Provider::Package' refers to the wrong object. An explicit
'require puppet/provider/package' overwrites an earlier 'Puppet::Provider::Package'
constant, resulting from a mixin of Puppet::Util and Puppet::Provider.