(PUP-9818) Ensure env name is always a symbol
Previously the StaticDirectory environment type did not aggressively
symbolize the environment name before comparing it, unlike other
environment types. This caused issues when some code paths provided the
name as a symbol when trying to load the config, because the requested
name did not match the saved name.

This also fixes the config lookup to provide an empty array rather than
an empty string for the global module path.