(PUP-9722) File mode change on every puppet run
Files present in a system protected filder, like C:\Windows or
C:\Program Files will inherit some additional ACE(Access Control Entries) from the
protected dir. So the file ACL(Access Control List) that stores ACEs will have
additional entries.

The ACE added is "APPLICATION PACKAGE AUTHORITY\ALL APPLICATION PACKAGES"
with a SID(Security Identifier) of "S-1-15-2-1".
"S-1-15-2-1” is not identified by security.rb as a well known SID,
as it was added in windows 8 to the list of well known SIDs
(https://msdn.microsoft.com/en-us/library/cc980032.aspx), and it is used for
execution of sandboxed applications.

Since it was unable to match to a well known SID, it was unable to map the SID
into POSIX mode, so when getting the mode, it added the S_IEXTRA bit flag indicating
that an access control entry was found whose SID is neither the owner, group, or other.

After the SID was added to the list of well known SIDs, it won't add the S_IEXTRA
bit when getting the file mode, so the desired mode matches the actual
mode and there is no need to apply the mode on every run.