(PUP-9816) Fix `puppet parser validate` handling of STDIN
In PUP-8984 the `puppet parser validate` face was refactored to
support the `--render-as json` option and also support validating
multiple files with parse errors in a single execution. (Previously it
would exit immediately on the first parse error.)

The implementation of PUP-8984 introduced a data structure to collect each
file that contained a parse error as well as the error itself.
Unfortunately a bug was introduced where when a manifest with no parse
errors was passed to `puppet parser validate` via the STDIN pipe (e.g.
`puppet parser validate < good.pp`) an entry in the "files with errors"
data structure was still created for "STDIN" but with a value of `nil`
instead of a parse error object. This empty entry caused the output renderer
to pass `nil` to the formatting methods resulting in `Error: undefined method
`message' for nil:NilClass`.

The fix for this issue adds a conditional so that the error collection
data structure entry for "STDIN" is only created if an actual parse
error was encountered, which mirrors the behavior when parsing manifests
from filename arguments.

A new unit test is also added to cover this scenario.