(PUP-8736) Symlink puppet device certs to /etc
Prior to this commit, SSL certificates, along with the rest of the
Puppet device configuration was stored under cache.

When the proxy Puppet agent for a device is the master, and the user
executes a backup and restore (manually, or as automated in OpsWorks),
the certificate and keys in the `ssldir` in each `devicedir` are lost,
requiring certificate regeneration for those devices.

This commit symlinks the `ssl` directory from
`/opt/puppetlabs/puppet/cache/devices/DEVICE/ssl` to
`/etc/puppetlabs/puppet/devices/DEVICE/ssl`, ensuring that SSL
certificates are not lost when cache is deleted.