(maint) Skip ssl tests when running on a single node
Normally, tests tagged with `server` are run with a layout of at least
two hosts, a master with an agent and a plain agent. However, they
shouldn't fail when run with only one node. These ssl tests are not
meaningful with only a master node, so this commit updates them to be
skipped when running with only one node.