(PUP-9909) Prefer SHA-256 for forge downloads
The forge API recently added `file_sha256` to the fetch module release
response[1] containing the SHA-256 checksum of the module release tarball. This
commit modifies the `puppet module install` command to prefer SHA-256, or
fallback to MD5 as it did before. If the PMT falls back to MD5 and FIPS is
enabled, then raise an exception like we did before. If both checksums are
missing, raise an exception.

We cannot drop MD5 yet, because some 3rd party forge implementations like
Artifactory haven't been updated to respond with `file_sha256`.

[1] https://forgeapi.puppetlabs.com/#operation/getRelease