(PUP-9513) Deprecate ssl_{server,client}_ca_auth settings
The `ssl_server_ca_auth` setting is dead and has been since the ruby webrick
puppet master was removed. Any attempt to get or set the setting will generate a
deprecation warning.

The `ssl_client_ca_auth` setting was created in 3.x to support certificate
chains back when `ca.pem` could only contain a single root CA. The setting is
read whenever `Puppet::Network::HttpPool.http*instance` is called. To prevent
deprecation spamming, only warn if the value is set, as that may indicate the
`ca.pem` does not contain all of the CA certificates that the agent needs to
verify the server's SSL chain.