(PUP-8949) use install_options in the 'latest' methods
Prior to this commit ...

The 'latest_with_new_pip' and 'latest_with_old_pip methods' (called by the
'latest' method) failed to execute `pip install` with the resource's
`install_options` attribute, potentially resulting in a lack of idempotency.

With this commit ...

Any 'install_options' are applied whenever 'install' is executed.