(PUP-9179) Acceptance test
Acceptance test to verify that if first puppet run runs as system, the
files in the cache directory have the correct group permissions.

This test simulates the behaviour as noted in the ticket by clearing the
state and catalog files and scheduling a puppet agent run as SYSTEM.
Assuming the files are created with the correct permissions, a "normal"
subsequent puppet run should be successful.