(PUP-9927) Don't call OpenSSL::PKey.read
The method is broken in MRI 2.3, doesn't exist in JRuby 9.1, and is broken in
JRuby 9.2. Just call the appropriate key subclass based on the PEM header. Also
remove our monkey patch since it's no longer used.

As a result of this change, the exception message is different for different
ruby versions and platforms when reading an EC password protected private key
with the wrong password.