(PUP-9930) Conditionally load openssl when not in JRuby FIPS
JRuby ships with a version of OpenSSL that is not FIPS compliant and is
its dependencies are stripped out of FIPS compliant builds of Puppet
Server. This causes loading "openssl" from JRuby to crash the server.

To work around this issue we create a wrapper file around loading
openssl or net/https (which depends on openssl). If the platform is
JRuby and is FIPS enabled. We must define several constants that are
wrapped directly by the Indirector even though they are not used server
side.