(PUP-9942) Allow puppet to skip a proxy
Commit 90560edf0f was supposed to allow puppet to skip a proxy if the host
we're trying to connect to matched an entry in the `NO_PROXY` environment
variable.  However, it didn't actually work, because it passed the `Site`
object instead of a String or URI to `no_proxy?`.

Pass a string representation of the site URI and update tests to not stub
`no_proxy?` behavior.