(PUP-9719) Additional modes for internal files
Add finer grade "mode" handling to cover internal puppet files e.g.
catalog and last_run files in addition to the SSL files that this
method was originally intended for.

Also fixup errors with mode 440, 444

Add a "health" warning comment indicating that this method should really
only be used for internal files and not the "file:" provider.