(PUP-5934) Add optional fact refresh after applying catalog
This commit adds a `resubmit_facts` option to Puppet's settings.
When set to `true`, this option causes the agent to reload facts
after applying a catalog and then submit the updated data to the
Puppet Server. This has the following benefits:

  - Changes in facts due to resources that were enforced during
    the run are reported immediately.

  - Agents that are running off cached catalogs can still submit
    regular updates to facts.

Note that enabling `resubmit_facts` will increase the load on
backend services such as Puppet Server and PuppetDB and is
therefore a change that should be monitored closely in
large environments.