(PA-2806) Pluginsync before running facts upload
When executing puppet facts upload, custom facts or plugin facts are not
syncronized with the master node resulting in missing facts or facts
that are not up to date.

This fix executes a plugin sync before the puppet facts upload is run.
Also temporary changes the log level to display info messages.