(PUP-9973) Always stringify execute array arguments
Previously, calling execute with an array of arguments and `sensitive: true`
options would fail to stringify the array. If one of the arguments wasn't a
string, such as the useradd provider passing a `uid`, then the call to
Kernel.exec would generate a "no implicit conversion of Integer into String"
error.

Change the redaction logic so that we always process the `command` string or
argument as before, and afterwards decide whether to redact the command
displayed in debug output.