(PUP-7808) Serialize valid YAML
The Puppet string serialization produce invalid YAML strings where
quotes are found.  Add a unit test that ensures the produced YAML is
valid and does not raise an exception when parsed.

Rely on YAML#dump instead of Puppet's function for serializing objects.
While this change the raw YAML representation of resources, it produce a
valid YAML document that can be used without issues.