(PUP-9994) implement $trusted['external'] facts
These facts are supplied through a new external lookup. They can be used
for injecting additional facts from third-party databases like asset
management or service-now.

To configure a lookup, provide the name of your lookup command to
`trusted_external_command`. The lookup script will receive the certname
as first and only argument and is expected to return the external facts
as a JSON hash on stdout.