(PUP-9991) Allow module uninstall in FIPS
The `uninstall` action detects if there are local modifications to a module
before deleting it. Since module checksum metadata is hardcoded to use MD5, the
entire uninstall action was disallowed in FIPS mode.

This commit relaxes the restriction so that uninstall is allowed provided
`--ignore_changes` or `--force` are specified, since they both skip the check
for local modifications.