(PUP-9990) Use Net::HTTP.new instead of Net::HTTP::Proxy
Previously puppet would attempt to connect via a proxy if the HTTP_PROXY
environment variable was set, but the destination server matched the
Puppet[:no_proxy] setting. The confusion occurred because ruby's Net::HTTP
automagically resolved proxy settings from the environment, after puppet had
already decided not to use one.

The disconnect is due to ruby 2.0 changing the default behavior of
Net::HTTP.new, which is called by Net::HTTP::Proxy[1]. If the third argument to
the initialize method (`p_addr`) is not specified, then it defaults to :ENV
causing ruby to lookup proxy configuration from the environment. In ruby 1.9 and
before, `p_addr` defaulted to nil.

Net::HTTP::Proxy is also marked as obsolete, so this commit switches to
Net::HTTP.new instead. If puppet decides not to use a proxy, it explicitly
passes nil arguments.

[1] https://github.com/ruby/ruby/commit/9416becda4592f7884460a054fe12b7761f6f055