(PUP-9931) Use ruby threadlocal implementation
If we cycle out JRuby instances while using the Java ThreadLocal class,
we end up retaining references to the old JRuby instances. A simple
solution to this is to use the native ruby ThreadLocalVar
implementation, so that when a JRuby instance is GCd everything
stored on the ruby Thread class will be collected as well.