(maint) Always stub state machine (#7707)
Commit b375c9792e added a new describe block in puppet#5.5.x, and that
version doesn't contain the SSL state machine. When the commit was merged
up to 6.4.x, it causes the state machine to run, causing the test to hang
indefinitely.

This commit moves the state machine stub to the top level `before` block so that it applies to all tests (now and in the future).