(PUP-10142) Move variable settings to context
Several settings are routinely updated before invoking a Puppet
subsystem as a way to pass inputs into said subsystem. In this way these
settings act essentially as global variables and consequently are not
threadsafe.

This moves the settings that change to the context and updates
subsystems that reference them to look in the context which is already
threadsafe.

These settings are notably `tasks`, `strict`, `strict_variables`, &
`code` and the subsystems they interact with are the parser & compiler.