(PUP-6382) Additions to Windows Delayed Auto Start
This commit updates the `enable` property for the service resource to
accept the `delayed` value. When set it will mark the specified Windows
service as a `Automatic (Delayed Start)` service.

This functionality requires implementation of the
`ChangeServiceConfig2W` and `QueryServiceConfig2W` functions.

Based on the Windows documentation, any service can be marked as a
delayed auto-start service, but the setting has no effect unless the
service is an auto-start service. Hence, when the `enable` property is
set to `delayed`, the service is first set to auto-start if it's
disabled.

Further reference:
https://docs.microsoft.com/en-us/windows/win32/api/winsvc/nf-winsvc-changeserviceconfig2w
https://docs.microsoft.com/en-us/windows/win32/api/winsvc/nf-winsvc-queryserviceconfig2w
https://docs.microsoft.com/en-us/windows/win32/api/winsvc/ns-winsvc-service_delayed_auto_start_info
https://sqlsanctum.wordpress.com/2015/05/06/remotely-get-set-delayed-start-for-a-service-in-powershell/