(PUP-10053) Only enable VCR for specific tests
Previously, if a test stubbed a request using webmock, but the stub did not
match, then vcr would intercept the attempted connection, raise
VCR::Errors::UnhandledHTTPRequestError, and print a confusing error about the
test missing a cassette.

The "register OID" test wasn't stubbing the request properly, but the problem
was masked by the code rescuing the UnhandledHTTPRequestError message, and
wrapping it in Puppet::Error, whose message matched `/./`.

With this change, VCR is disabled by default, and only enabled for tests marked
with `:vcr => true`.

If the test isn't using vcr, then unhandled requests will be intercepted by
webmock, and more useful error will be displayed describing how to stub the
request.