(PUP-9978) Support for DNF modules
This commit introduces support for managing DNF/YUM modules through the
`module` subcommand. Modules are groups of packages that represent an
application, a language runtime, or any logical group.

Modules can be available in multiple streams, usually representing a
major version of the software they include. Profiles are package subsets
representing a specific use case of the module (these are handled by the
`flavor` parameter of the package type).

A sample module installation using all applicable flags would be:

```
package { 'postgresql':
  ensure   => '9.6' # stream
  flavor   => 'client' # profile
  provider => 'dnfmodule'
}
```

Due to the significant difference between a package and a module,
`dnfmodule` is an opt-in provider and should be explicitly specified in
the manifest.