(PUP-9647) Quote pip path when calling execpipe
If the resolved pip path contains a space, then calls to `execute` will work,
but `execpipe` will not, because `execute` calls into `CreateProcess` which
handles a command with spaces, while `execpipe` calls `Kernel.open` which does
not.

If the path contains a space, wrap in quotes before calling `execpipe`. This
change is less risky then switching everything to `execute` due to the different
ways that they handle environment variables like `LC_ALL`.

Thanks to Sean Millichamp <sean@bruenor.org> for identifying why this
was failing on Windows.