(maint) Use Ruby 2 style caller()
In Ruby 2, and JRuby 9k, `Kernel#caller` got a new optional argument that
controls how many frames to go back when constructing the call stack.

In most cases where we call `caller` we also throw away most of the
call stack. This has performance implications for every Ruby environment
and especially for JRuby.

Most of the time this doesn't matter because we are getting the call
stack for error reporting, however in at least one case this
information is used for non-fatal warnings.