(PUP-10080) Stop Exec defaulting its cwd parameter
Without this patch, the base Exec provider defaults the cwd parameter to
Puppet's working directory.  As a result, in the event that no cwd is
explicitly specified, Puppet attempts to change to the directory that is
already its working directory when it runs the Exec's command.  In most
cases that's merely wasteful, but under some circumstances it causes the
Exec to fail needlessly.

With the defaulting removed, the provider's
Puppet::Util::Execution.execute() call just does the right thing.
That is, it avoids attempting a working-directory change in the event
that no value was specified for the cwd parameter.