(PUP-9562) Respect forcelocal for user comment
A user resource configured with forcelocal will still try to sync the
comment with the external directory services. It does use `lusermod` to
modify the local `/etc/passwd` to the comment specified in the user
resource, but it compares the `in_sync` with the external directory
services, meaning that it always updates the comment on catalog
compilation.

This commit queries the comment from the local `/etc/passwd` if the
`forcelocal` parameter is set. It also modifies the `finduser` method to
return a hash of parameter/value type, since returning numbered elements
is more error-prone.