(PUP-1537) add hold property for package
This commit adds `hold` property for package providers on
Debian(dpkg/apt/fink/aptitude) and Solaris(pgk) as
alternative of using "held" value for `ensure`.

Using "held" value for `ensure` works as before, but it shows
deprecation warning and will be removed in further release.

Allowed values for `hold` are "yes"/"true"/"no"/"false",
default to "false". Hold can be specified with or without `ensure`,
if `ensure` is missing will default to "present".

Hold cannot be specified together with "purged", "absent" or "held"
values for `ensure`.