(PUP-1537) add mark property for package
This commit adds `mark` property for package providers on
Debian(dpkg/apt/fink/aptitude) and Solaris(pgk) as
alternative of using "held" value for `ensure`.

Using "held" value for `ensure` works as before, but it shows
deprecation warning and will be removed in further release.

Allowed values for `mark` are "hold/none", default to "none".
Mark can be specified with or without `ensure`, if `ensure` is
missing it will default to "present".

Mark cannot be specified together with "purged", "absent" or "held"
values for `ensure`.