(PUP-9562) useradd provider improvements
This commit changes the `passwd_keys` array elements to be symbols
instead of strings (since we do not manipulate them in the code), and
updates the tests accordingly. Also, we do not explictly close the
file since Ruby should do it for us when leaving the block's context.

This commit also adds tests for `useradd#finduser` and fixes up some
wrong expectations.