(PUP-7218) systemd service provider missing double dash
Without applying this page any service beging with a '-' (dash) cannot
be started by systemd. This is a problem because it's conventional in
systemd services to do so. This patch allows users to run commands
systemctl enable [service] and systemctl disable [service] by
appending a -- (em dash) between the command and service name. Tests
have been updated to verify corrected output.