(PUP-10021) fix net user /add prompt in tests
Windows has a password policy, where it prompts the
user based on password complexity, password length
and history. In our case, it is prompting since password
exceeds the limit of 14 characters. This is to enable
using the user accounts on Windows versions older
than Windows 2000, since they don't support long passwords.

Since we are using `net user /add` only in tests,
it is safe to use a shorter password for this test.