(PUP-10102) fix service crash
When running on distributions that do not have a
default service provider puppet will try to
evaluate `runit` and set `defpath` default
before verifying that `runit` is suitable. Raising an
error when setting defpath results in a state that should
not be reached.

This commit removes the raise exception from defpath
and let puppet decide if the provide is suitable or not.