(PUP-10113) Speed up method calls and calls to []
Before this, the logic that evaluates method calls and the logic
that evaluates the [] operator would do a type inference and then check
if the type supports methods.
This was bad because it requires a full type inference that may be quite
expensive (for a large hash for example).

This is now fixed by adding a specific inference method in the
TypeCalculator that quickly returns nil for all types known to not have
methods, and infers the type for the rest.