(PUP-10106) Don't require domain proxy exceptions to start with dot
Previously, the `no_proxy` setting only supported domain exceptions using a
wildcard (*.example.com) or a dotted-domain (.example.com). However, curl and
ruby accept a dotted-domain and a domain suffix (example.com, ample.com), but
not wildcards.

This could lead to strange behavior if puppet executed ruby or curl during an
exec resource, and the `no_proxy` environment variable was specified as a domain
suffix -- puppet would still use the proxy, while ruby & curl would not.

This commit allows puppet to accept `no_proxy` environment variables or puppet
settings containing domain suffixes, while retaining support for wildcards and
dotted domains. As a result, it's no longer possible to force puppet to not use
a proxy for a specific host, but not any domains that happen to match that
hostname.