(PUP-10033) Retry after if service is unavailable
If we receive an HTTP response 429 or 503, then parse the `Retry-After` header,
which can either be some positive number of seconds or an RFC 2822 date. If the
header is present but cannot be parsed, raise an error. If the header is
missing, retry immediately. If the header is present and can be parsed, wait
that many seconds or time in the future, before retrying the request.

But don't wait more than the `runinterval`, as that could lead to overlapping
agent runs, when puppet is run from cron.

Allow puppet to retry up to 100 times, so we don't get into an infinite retry
loop.