(PUP-10057) Windows confuses domain/local accounts
Behavior of the user resource goes wonky when an AD account exists with
the same name as the local user account you’re trying to manage on a
domain-joined Windows server.

This commit queries the WMI database for the computer's Domain Role. If
it's domain-joined, it will compare the found user's domain with the
machine hostname to confirm the account exists locally.

This issue is caused by the LookupAccountNameW function automatically
trying to resolve the name using domain controllers trusted by the local
system if the name cannot be resolved on the local system.