(PUP-10121) Simplify connection error handling
Rather than create a new exception class when we catch connection
errors, we should raise the original exception. Previously, this rescue
block was modified to add additional contextual information to the
exceptions we were catching to give the user more information around the
context of failures. Rather than raising the original message, it would
create a new exception with the additional information.

This commit simply logs the error with Puppet with the additional
information, and raises the original caught exception. This saves us
from errors that stem from different error initializer expectations.