(PUP-10040) Use route URL instead of assuming :ca_server setting
Since the route to the CA may be due to SRV, print the URL from the route
instead of the hard coded settings.

Also don't emit 'Downloaded certificate' twice in `puppet ssl`, and add a
missing `Downloaded certificate` in the state machine, as we're no longer
using `Puppet::Rest::Routes`.