(maint) Rename Route to ServiceType
The term 'route' was a holdover from an earlier version of the code when
`route_to` returned a Route object. But that was changed so that `route_to`
returns a connected `Service` where the service contains the URL representing
the route to a specific host and port.

As a result, rename `Route` to `ServiceType` as the latter describes the
information needed to create each type of service.