(PUP-10149) Fix regression in systemd daemon-reload
This commit restores the functionality of
https://tickets.puppetlabs.com/browse/PUP-3483

It was broken my the merge-up of the PR for
https://tickets.puppetlabs.com/browse/PUP-7218

Without this change

```
systemctl show -- some.service --property=NeedDaemonReload
```

shows *all* properties and the `.strip.split('=').last` doesn't work.
The option `--property=NeedDaemonReload` is not being treated as an
option, but a very obscurely named service.

The fix is trivial.  By moving the `--property=NeedDaemonReload` to be
before the `--`, it works again.