(maint) Add a rake task to run tests with docker
This uses git-based acceptance, skips setting up a master, and does not
run any tests tagged `server`. It also creates and munges a hosts file
for beaker to use. This removes any manual munging that was otherwise
needed when running acceptance tests with docker.

This commit also modifies the git testing to prevent the automation from
provisioning a server. This might be too big of a hammer for this, but
generally git-based acceptance tests should be light weight. The tests
should error if we ever run one that does require a master, but
generally this should be safe.