(PUP-10039) Add ServerList resolver
This commit adds a resolver that will iterate through an array of
servers and return the first in the list that responds successfully to a
get connection. We do not yet use this new resolver anywhere, but will
switch over to it in the near future.

If a port is not provided in the list of servers, it will use a default
port. If no functional server is found, we raise an error rather than
allowing the session to continue searching other resolvers for a
functional server.