(maint) Create and check service in the resolver
Rather than checking a connection is successful as a part of the
session, this commit changes to let us check that connection in the
resolvers. This is something we needed to do for ServerList, so this
change makes that behavior consistent for all other resolvers as well.
It also allows us to reduce some complexity. We can now create the
service directly from the resolver, rather than having to pass that call
down from the resolver, to the session, and finally to the service.