(maint) Reset "tasks" setting after with_script_compiler
The `Puppet::Pal.with_script_compiler` method sets the
`Puppet[:tasks]` setting to true for the duration of the block, but
never sets it back to its original state. This means that the mode of
operation the process is in will be globally changed depending on
whether this method has been called. The corresponding
`with_catalog_compiler` method has similar behavior in setting
`Puppet[:tasks]` to false, but it ensures that the setting is reset to
its previous value when the method returns. We now copy that behavior in
`with_script_compiler`.