(PUP-10157) Observe ServerList Resolver
This commit enables the ServerList resolver for the ca service and adds
tests to ensure a consistent flow when resolving the route to the
server.

This commit also does a bit to enforce the resolver policy for different
services. For example, we should only use `:server_list` when resolving
for the :puppet service generally. When we are resolving the :ca
service and we have not explicitly set `:ca_server`, then we may also
use `:server_list`. If we do traverse the list of servers and do not
find a functional one, we should fail. Otherwise, we should return nil
so we can try the following resolvers in our list.

Generally, we're trying to keep policy out of the resolvers, but in an
effort to maintain consistent behavior, we had to make some exceptions.

This commit also adds more tests to define the order in which we try
different resolvers, and what happens when we do or do not find a
functional server with a given combination therein.

We also have to modify the existing ServerList tests to pass in the new
required variables