(maint) Use let(:facts) instead of @facts
Ensure that we have an empty set of "saved" facts in memory. Use let instead of
@facts.

Don't stub `facts_for_uploading` as it turns out during `puppet apply`
we collect facts as a side effect of creating the node. But then we don't send
them in the catalog find request. The compiler then calls facter (again).
Ironically the test for this was broken and didn't assert that the facts were
in-fact empty.

Update URL encoded facts test to match reality.