(maint) Add tests for cached catalog behaviors
The cached catalog is updated as a side-effect of calling Catalog.indirection.find,
yet that method was stubbed.

Use webmock to stub the rest request, and verify the cache is updated (or not
in the case of noop mode). Also verify `apply` mode doesn't update the cached
catalog, and that the run succeeds when the cache is nil.