(PUP-10160) Don't conflate cached and ral catalogs
Previously, `options[:catalog]` could refer to the cached RAL catalog or the
RAL catalog passed in by `puppet apply`. The code deleted the catalog from the
options hash, because if the catalog's environment didn't match, then we didn't
want to retry the cached catalog again.

Rather than conflate the puppet apply and cached catalog cases, keep track of
them separately so that we don't mutate the options hash during the configurer's
run.