(maint) Create man directories before manpage generation
Running `mkdir -p man/man{5,8}` makes `/bin/sh` (which is the default
shell for rake) create a single directory titled `man{5,8}`, since this
type of grouping is bash-specific.

The previous implementation might work on most modern systems where
`/bin/sh` is a symlink to `bash`, but to be safe we create the
directories individually.