(PUP-7113) Document AIX local resource functionality
To manage a local user/group resource on AIX, the `ia_load_module`
parameter needs to be set to `files`. This has the same effect as
`forcelocal` on Linux.

Since the `forcelocal` parameter is linked to the `libuser` feature
which is platform-specific, we cannot use it on AIX.

Extend the documentation on these flags to reflect this use case.