(PUP-10160) Cached catalog takes precedence over options catalog
In 5.5.17 and prior, if `use_cached_catalog` was set and we loaded the cached
catalog, then the cached catalog overwrite the catalog in the options hash[1].

In commit f490127b4554a2bb4d30117c8f581cb7cb816a7f, we no longer conflated the
cached and options catalogs, but the precedence was flipped.

This commit ensures the cached catalog takes precedence as it did before and
adds tests.

It also removes some unnecessary preconditions from some tests.

[1] https://github.com/puppetlabs/puppet/blob/5.5.17/lib/puppet/configurer.rb#L255