(PUP-10165) Allow hostprivkey and hostcert to override certname
Previously the hostprivkey and hostcert settings were ignored, and paths to the
client cert and private key were resolved based on the cert/key directories and
certname setting.

This is acceptable for puppet agents, because puppetserver generally requires
that the client cert's common name (and therefore <name>.pem) match the node's
identity in the HTTP request (which is determined by certname).

But this is an issue for other parts of the puppet ecosystem, such as running
`puppet apply` and using a trusted client cert to connect to puppetdb.

To allow this, allow the hostprivkey and hostcert puppet settings to override
the default paths.