(PUP-10150) Explicitly print PuppetStack when appropriate
Prior to a refactor of how PuppetStack worked we would print the Puppet
stack frames as part of printing the Ruby stack. This was desirable in
some cases and in others it's believed that just printing the Puppet
stack would be best.

This patch re-adds the interleaving of the Puppet stack with the Ruby
stack but as an explict action during exception printing vs as a side
effect of including the Puppet stack as part of the Ruby stack.

This is achieved by adding a new accessor, `puppetstack` to our Errors
that deal with the reading of files. When an exception of this kind is
created it saves a reference to the current PuppetStack. Exceptions that
are created from these inherit the original exception's puppetstack.
Puppet::Util::Logging.build_exception_trace (primarily builds traces as
arrays for ParseErrorWithIssues),
Puppet::Util::Logging.format_exception (formats traces for
non-ParseErrorWithIssues as strings), and their underlying usage of
Puppet::Util.pretty_backtrace have been refactored to honor the
puppetstack if configured and available on the exception in question.

It also adds a new setting `:puppet_trace` which will only print the
Puppet stack, without any of the Ruby stack frames for those authors
working solely on Puppet code.