(PUP-10171) Support changing package flavors
With providers where we have the `flavor` parameter, we cannot change it
once a package is installed. The only way to do it is to use `ensure =>
absent` on the package, then install the different flavor.

This behavior is achievable on both `dnfmodule` and `openbsd` package
providers. To do this we change `flavor` to be a property instead of a
parameter, so it gets tracked as part of the resource state.

To ensure we don't call flavor methods on unsupported providers, we
confine the property to the newly added `supports_flavors` feature.