(PUP-10142) Proxy variable settings calls to context
This patch updates Settings#[] and Settings#[]= to check if the setting
in question was moved to the context and if so lookup/set it there. It
will also issue deprecation warnings for the attempt. It provides a new
method, `get_parsed_setting_for_context` to allow those settings to be
read without deprecation warnings when bootstrapping the context.